Доработан контроллер, сервисы, добавлена обработка ошибок, фабрика DTO, доработан UI.

+++ src/main/java/ru/service/shelter/services/AnimalsServiceImpl.java
@@ -3,10 +3,11 @@ package ru.service.shelter.services;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
-import ru.service.shelter.dto.AnimalsDTO;
+import ru.service.shelter.dto.FactoryAnimalDTO;
 import ru.service.shelter.dto.NewAnimal;
+import ru.service.shelter.entity.Animal;
 import ru.service.shelter.entity.AnimalGender;
-import ru.service.shelter.entity.Animals;
+import ru.service.shelter.exseption.ResourceNotFoundException;
 import ru.service.shelter.repositories.AnimalTypeRepository;
 import ru.service.shelter.repositories.AnimalsRepository;
 
@@ -14,13 +15,19 @@ import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.Optional;
 
 @Component
 @Transactional
 public class AnimalsServiceImpl implements AnimalService{
     AnimalsRepository animalsRepository;
     AnimalTypeRepository animalTypeRepository;
+    FactoryAnimalDTO factoryAnimalDTO;
+
+    @Autowired
+    public void setFactoryAnimalDTO(FactoryAnimalDTO factoryAnimalDTO) {
+        this.factoryAnimalDTO = factoryAnimalDTO;
+    }
 
     @Autowired
     public void setAnimalsRepository(AnimalsRepository animalsRepository) {
@@ -33,40 +40,49 @@ public class AnimalsServiceImpl implements AnimalService{
     }
 
     @Override
-    public List<AnimalsDTO> getAllAnimals() {
-        List<AnimalsDTO> animalsList = animalsRepository.findAll().stream()
-                .map(animals -> new AnimalsDTO(animals))
-                .collect(Collectors.toList());
-        return animalsList;
+    public List<Animal> getAllAnimals() {
+        return animalsRepository.findAll();
     }
 
     @Override
-    public Animals getAnimalById(long id) {
-        return animalsRepository.getById(id);
+    public Animal getAnimalById(long id) throws ResourceNotFoundException {
+        Optional<Animal> animal = Optional.of(animalsRepository.getById(id));
+        if(!animal.isPresent()){
+            throw new ResourceNotFoundException("Запись не найдена в БД");
+        }
+        return animal.get();
     }
 
     @Override
-    public void deleteAnimalsById(long id) {
+    public Animal deleteAnimalById(long id) throws ResourceNotFoundException {
+        Optional<Animal> animal = Optional.of(animalsRepository.getById(id));
+        if(!animal.isPresent()){
+            throw new ResourceNotFoundException("Запись не найдена в БД");
+        }
         animalsRepository.deleteById(id);
+        return animal.get();
     }
 
     @Override
-    public void updateAnimals(NewAnimal newAnimal) {
-        Animals animal = animalsRepository.getById(newAnimal.getId());
-        animal.setAnimalType(animalTypeRepository.getById(Long.parseLong(newAnimal.getAnimalType())));
-        animal.setAnimalGender(Enum.valueOf(AnimalGender.class, newAnimal.getAnimalGender()));
+    public void updateAnimal(NewAnimal newAnimal) throws ResourceNotFoundException {
+        Optional<Animal> animal = Optional.of(animalsRepository.getById(newAnimal.getId()));
+        if(!animal.isPresent()){
+            throw new ResourceNotFoundException("Запись не найдена в БД");
+        }
+        animal.get().setAnimalType(animalTypeRepository.getById(Long.parseLong(newAnimal.getAnimalType())));
+        animal.get().setAnimalGender(Enum.valueOf(AnimalGender.class, newAnimal.getAnimalGender()));
         try {
-            animal.setDateOfReception(getDate(newAnimal.getDateOfReception()));
+            animal.get().setDateOfReception(getDate(newAnimal.getDateOfReception()));
         } catch (ParseException e) {
             e.printStackTrace();
         }
-        animal.setDescription(newAnimal.getDescription());
-        animalsRepository.saveAndFlush(animal);
+        animal.get().setDescription(newAnimal.getDescription());
+        animalsRepository.saveAndFlush(animal.get());
     }
 
     @Override
     public void addNewAnimal(NewAnimal newAnimal) {
-        Animals animal = new Animals();
+        Animal animal = new Animal();
         animal.setAnimalType(animalTypeRepository.getById(Long.parseLong(newAnimal.getAnimalType())));
         animal.setAnimalGender(Enum.valueOf(AnimalGender.class, newAnimal.getAnimalGender()));
         try {
@@ -79,8 +95,7 @@ public class AnimalsServiceImpl implements AnimalService{
     }
 
     private Date getDate(String day) throws ParseException {
-        if(day.isEmpty()|| day==null){return  new Date();}//отсекаем пустую дату
-        //переводим дату в нужный формат
+        if(day.isEmpty() || day==null) {return  new Date();}//отсекаем пустую дату
         SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
         return sdf.parse(day);
     }
